+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add `Unit.transientCache` and `Unit.getCopy`
 - Fix `ElementBondIterator` indices mapping logic for inter-unit bonds
 - Fix `pickPadding` and `pickScale` not updating `PickHelper`
+- MolViewSpec extension: support loading extensions when loading multistate files
 - Do not add bonds for pairs of residues that have a `struct_conn` entry
 - Improved `ma_qa_metric` support
   - Parse all local metrics